fix(slider): optimize
